Entryのクラスを共通にする
ジェネリクスをキレイに推論してくれるのが良き
